requirements.txt
flask
google-generativeai
pygame
gtts
python-dotenv
SpeechRecognition
pyaudio
scipy